+++ app/models/unfollowership.rb
@@ -21,7 +21,7 @@ class Unfollowership < ActiveRecord::Base
     unfollowerships = follower_uids.map.with_index { |follower_uid, i| [from_uid.to_i, follower_uid.to_i, i] }
 
     ActiveRecord::Base.transaction do
-      delete_all(from_uid: from_uid)
+      delete_all(from_uid: from_uid) if exists?(from_uid: from_uid)
       import(%i(from_uid follower_uid sequence), unfollowerships, validate: false, timestamps: false)
     end
   end
